Add MSYS2 launcher
Also makes mintty pinnable to Windows Taskbar.  Before, a pinned
shortcut would have generic icon and fail to both launch the shell as a
login one (so that it sources /etc/profile) and possibly set MSYSTEM to
indicate MinGW shell type.

The launcher will pop up on start and new-window with Alt-F2, when no
shell command to run in the terminal was passed on command line.  This
allows for parameters to be passed from shorcuts or batch files, while
still having the select-shell dialog displayed.

The list of shells is harcoded.  An editable config file in /etc would
be nice (for e.g. "MSYS2 zsh" item), but this works for now.

Usability note:
If the taskbar cannot forget the mintty's default icon, rebooting helps
(at least, a non-fast-restart one).
If the taskbar cannot forget the mintty's default "Terminal" tooltip
text displayed upon mouse hover of the shorcut, deleting the

 [HKEY_CLASSES_ROOT\Local Settings\Software\Microsoft\Windows\Shell\MuiCache]
 "C:\\path\\to\\msys\\usr\\bin\\mintty.exe.FriendlyAppName"

registry value (or just the entire MuiCache directory) should help, at
least for the current Windows 10.